storage:: write to datastorage.

+++ utils/utils.cpp
@@ -11,6 +11,8 @@ std::list<boost::filesystem::path> utils::ls(const std::string& path){
 }
 
 bool utils::rm(const std::string&rm_path){
+    if(!boost::filesystem::exists(rm_path))
+        return true;
     boost::filesystem::path path_to_remove(rm_path);
     for (boost::filesystem::directory_iterator end_dir_it, it(path_to_remove); it!=end_dir_it; ++it) {
         if(!boost::filesystem::remove_all(it->path())){
